refactor: supprime l'ancien type ConseillersEntreprisesVariant
