+++ site/source/components/ConseillersEntreprisesButton.tsx
@@ -4,6 +4,7 @@ import { styled } from 'styled-components'
 
 import { Body, Button, Emoji, PopoverWithTrigger } from '@/design-system'
 
+import { type ConseillersEntreprisesVariant } from './ConseillersEntreprises/BoutonConseillersEntreprises'
 import Loader from './utils/Loader'
 
 const LazyIframe = lazy<ComponentType<{ src: string; onLoad: () => void }>>(
@@ -24,13 +25,6 @@ const Container = styled.div`
 const ButtonLabel = styled.span`
 	margin-left: 1rem;
 `
-
-export type ConseillersEntreprisesVariant =
-	| 'activite_partielle'
-	| 'recrutement'
-	| 'micro_entrepreneur'
-	| 'revenus_par_statut'
-	| 'professions_liberales'
 
 export const ConseillersEntreprisesButton = ({
 	variant,
